Extracted geo specific functions to a service
