+++ src/app/projects/services/autofill-project.service.spec.ts
@@ -5,16 +5,22 @@ import {AutofillProject} from '../models/autofill-project';
 import {Identifier} from '../models/europe-pmc-search';
 import {AutofillProjectService} from './autofill-project.service';
 import {IngestService} from "@shared/services/ingest.service";
-import {ViewContainerRef} from "@angular/core";
+import {AlertService} from "@shared/services/alert.service";
+import {Router} from "@angular/router";
+
 
 describe('AutofillProjectService', () => {
   let service: AutofillProjectService,
     httpTestingController: HttpTestingController,
     httpClient: HttpClient,
-    ingestSvc: jasmine.SpyObj<IngestService>;
+    ingestSvc: jasmine.SpyObj<IngestService>,
+    router: jasmine.SpyObj<Router>,
+    alertSvc: jasmine.SpyObj<AlertService>;
 
   beforeEach(() => {
     ingestSvc = jasmine.createSpyObj('IngestService',['queryProjects']);
+    alertSvc = jasmine.createSpyObj('AlertService',['error']);
+    router = jasmine.createSpyObj('Router',['navigate']);
     TestBed.configureTestingModule(
       {
         providers: [
@@ -22,6 +28,14 @@ describe('AutofillProjectService', () => {
           {
             provide: IngestService,
             useValue: ingestSvc
+          },
+          {
+            provide: AlertService,
+            useValue: alertSvc
+          },
+          {
+            provide: Router,
+            useValue: router
           },
         ],
         imports: [HttpClientTestingModule]
